ESModelをModelとRepositoryに分割
