+++ jobHunt/Domain/Model/ESModel.swift
@@ -1,0 +1,20 @@
+//
+//  ESModel.swift
+//  jobHunt
+//
+//  Created by 出口楓真 on 2022/10/16.
+//
+
+import Foundation
+
+struct ESModel {
+    
+    var name: String?
+    var deadline: Date?
+    var motivation: String?
+    var gakuchika: String?
+    var strongPoints: String?
+    var weakPoints: String?
+    var other: String?
+    var category: EventName?
+}
